revise #727: adjust package structure
