+++ sharding-core/src/main/java/io/shardingjdbc/core/optimizer/OptimizeEngineFactory.java
@@ -17,11 +17,13 @@
 
 package io.shardingjdbc.core.optimizer;
 
+import io.shardingjdbc.core.optimizer.insert.InsertOptimizeEngine;
+import io.shardingjdbc.core.optimizer.query.QueryOptimizeEngine;
 import io.shardingjdbc.core.parsing.parser.sql.SQLStatement;
 import io.shardingjdbc.core.parsing.parser.sql.dml.DMLStatement;
 import io.shardingjdbc.core.parsing.parser.sql.dml.insert.InsertStatement;
 import io.shardingjdbc.core.parsing.parser.sql.dql.select.SelectStatement;
-import io.shardingjdbc.core.routing.sharding.GeneratedKey;
+import io.shardingjdbc.core.routing.router.GeneratedKey;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
