+++ sharding-core/src/main/java/io/shardingjdbc/core/routing/router/GeneratedKey.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingjdbc.core.routing.sharding;
+package io.shardingjdbc.core.routing.router;
 
 import io.shardingjdbc.core.parsing.parser.context.condition.Column;
 import io.shardingjdbc.core.parsing.parser.context.condition.GeneratedKeyCondition;
