+++ sharding-core/src/test/java/io/shardingjdbc/core/optimizer/InsertOptimizeEngineTest.java
@@ -18,14 +18,15 @@
 package io.shardingjdbc.core.optimizer;
 
 import io.shardingjdbc.core.api.algorithm.sharding.ListShardingValue;
+import io.shardingjdbc.core.optimizer.condition.ShardingConditions;
+import io.shardingjdbc.core.optimizer.insert.InsertOptimizeEngine;
 import io.shardingjdbc.core.parsing.parser.context.condition.AndCondition;
 import io.shardingjdbc.core.parsing.parser.context.condition.Column;
 import io.shardingjdbc.core.parsing.parser.context.condition.Condition;
 import io.shardingjdbc.core.parsing.parser.context.condition.OrCondition;
 import io.shardingjdbc.core.parsing.parser.expression.SQLNumberExpression;
 import io.shardingjdbc.core.parsing.parser.expression.SQLPlaceholderExpression;
-import io.shardingjdbc.core.routing.sharding.GeneratedKey;
-import io.shardingjdbc.core.routing.sharding.ShardingConditions;
+import io.shardingjdbc.core.routing.router.GeneratedKey;
 import org.junit.Test;
 
 import java.util.Collections;
